+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 - Wilder Wild's blocks are now properly included in the `minecraft:replaceable` Blocks Tag.
 - Added the Stripped Poplar Log and Stripped Hollowed Poplar Log blocks.
 - Wilder Wild's Material Rules (formerly known as Surface Rules) are now data-driven.
+- Wilder Wild's custom Cactus damage sound now only plays if the Cactus block sounds config option is enabled.